items+effects: introduce manager modules
