+++ src/meson.build
@@ -208,7 +208,8 @@ common_sources = [
   'trx/game/creature/shooting.c',
   'trx/game/cutscene.c',
   'trx/game/demo.c',
-  'trx/game/effects.c',
+  'trx/game/effects/draw.c',
+  'trx/game/effects/manager.c',
   'trx/game/events.c',
   'trx/game/fader.c',
   'trx/game/fmv.c',
@@ -281,7 +282,8 @@ common_sources = [
   'trx/game/items/actions/lara.c',
   'trx/game/items/anim.c',
   'trx/game/items/col.c',
-  'trx/game/items/common.c',
+  'trx/game/items/draw.c',
+  'trx/game/items/manager.c',
   'trx/game/items/utils.c',
   'trx/game/items/walkable.c',
   'trx/game/lara/breath.c',
