+++ src/trx/game/effects.h
@@ -1,5 +1,6 @@
 #pragma once
 
-#include <trx/game/effects/common.h>
 #include <trx/game/effects/const.h>
+#include <trx/game/effects/draw.h>
+#include <trx/game/effects/manager.h>
 #include <trx/game/effects/types.h>
