+++ src/trx/game/effects/draw.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <stdint.h>
+
+void Effect_Draw(int16_t effect_num);
