+++ src/trx/game/items.h
@@ -3,9 +3,10 @@
 #include <trx/game/items/actions.h>
 #include <trx/game/items/anim.h>
 #include <trx/game/items/col.h>
-#include <trx/game/items/common.h>
 #include <trx/game/items/const.h>
+#include <trx/game/items/draw.h>
 #include <trx/game/items/enum.h>
+#include <trx/game/items/manager.h>
 #include <trx/game/items/types.h>
 #include <trx/game/items/utils.h>
 #include <trx/game/items/walkable.h>
