+++ src/trx/game/items/draw.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <trx/game/items/types.h>
+
+void Item_ControlDraw(ITEM *item);
