+++ src/trx/game/items/manager.h
@@ -21,7 +21,6 @@ int16_t Item_Spawn(const ITEM *item, OBJECT_ID obj_id);
 
 void Item_Initialise(int16_t item_num);
 void Item_Control(void);
-void Item_ControlDraw(ITEM *item);
 void Item_Kill(int16_t item_num);
 void Item_KillAllActive(void);
 void Item_RemoveActive(int16_t item_num);
@@ -31,7 +30,6 @@ void Item_AddActive(int16_t item_num);
 void Item_UpdateRoom(int16_t item_num, int16_t room_num);
 
 int32_t Item_GlobalReplace(OBJECT_ID src_obj_id, OBJECT_ID dst_obj_id);
-bool Item_IsTriggerActive(ITEM *item);
 
 // Set the name of the item, storing a copy of the provided string.
 // Returns false if the name is already used by another item.
